refactor: Question, Answer model categories 변경
- model 사용 코드 변경
- useQuestion, useCategory 로직 분리
- getCategoryName 적용
