+++ src/constants/Question.ts
@@ -1,1 +1,3 @@
+export const Positions = ["frontend", "backend"] as const;
+export type Position = (typeof Positions)[number];
 export const ALL_CATEGORIES = -1;
