+++ src/models/Answer.model.ts
@@ -1,9 +1,9 @@
 export interface Answer {
   id: number;
-  content: string;
   question: {
     id: number;
     title: string;
-    categories: string[];
+    categories: number[];
   };
+  content: string;
 }
